feat(LottoWinningStatistics): 당첨 통계 기능 추가
- LottoWinningStatistics: 당첨 통계 책임 전담
- LottoProfitCalculator의 수익률 계산 책임을 LottoWinningStatistics로 이전
- LottoProfitCalculator 클래스 삭제
